refactor(webhooks): isolate completion delivery
